support generation of cyclonedx in json format
