+++ lib/sbom/cyclonedx.ex
@@ -1,125 +1,34 @@
 defmodule SBoM.CycloneDX do
   @moduledoc """
-  Generate a CycloneDX SBoM in XML format.
+  Generate a CycloneDX SBoM in XML and JSON format.
   """
 
-  alias SBoM.License
+  alias SBoM.CycloneDX.{Json, Xml}
 
   @doc """
-  Generate a CycloneDX SBoM in XML format from the specified list of
+  Generate a CycloneDX SBoM in XML or in JSON format from the specified list of
   components. Returns an `iolist`, which may be written to a file or IO device,
   or converted to a String using `IO.iodata_to_binary/1`
 
-  If no serial number is specified a random UUID is generated.
+  ### Options
+
+  :format - specifies the format given. If format is ':json', CycloneDX SBoM in JSON format is generated  otherwise
+            will be generated in XML format.
+
+  :schema - specifies the schema version used i.e '1.1', '1.2'. CycloneDX SBoM in XML will be generated based 
+            on the schema version passed. The default '1.1' schema version will be used if its not specified.
+
+  :serial - specifies the serial number used. If no serial number is specified a random UUID is generated.
   """
 
   def bom(components, options \\ []) do
-    bom =
-      case options[:schema] do
-        "1.1" ->
-          {:bom,
-           [
-             serialNumber: options[:serial] || uuid(),
-             xmlns: "http://cyclonedx.org/schema/bom/1.1"
-           ], [{:components, [], Enum.map(components, &component/1)}]}
-
-        _ ->
-          {:bom,
-           [
-             serialNumber: options[:serial] || uuid(),
-             xmlns: "http://cyclonedx.org/schema/bom/1.2"
-           ],
-           [
-             {:metadata, [],
-              [
-                {:timestamp, [], [[DateTime.utc_now() |> DateTime.to_iso8601()]]},
-                {:tools, [], [tool: [name: [["SBoM Mix task for Elixir"]]]]}
-              ]},
-             {:components, [], Enum.map(components, &component/1)}
-           ]}
-      end
-
-    :xmerl.export_simple([bom], :xmerl_xml)
-  end
-
-  def bom_json(components, opts) do
-    %{
-      bomFormat: "CycloneDX",
-      specVersion: options[:schema],
-      serialNumber: options[:serial] || uuid(),
-      version: "1",
-      metadata: _metadata(),
-      components: _components(components)
-    }
-  end
-
-  def _metadata do
-    %{
-      timestamp: DateTime.utc_now() |> DateTime.to_iso8601(),
-      tools: [%{name: "SBoM Mix task for Elixir"}]
-    }
-  end
-
-  defp _components(components) do
-    Enum.map(components, fn component ->
-      %{
-        type: component.type,
-        name: component.name,
-        version: component.version,
-        description: component.description,
-        licenses: Enum.map(component.licenses, fn l -> %{license: %{id: l}} end),
-        purl: component.purl
-      }
-    end)
-  end
-
-  defp component(component) do
-    {:component, [type: component.type], component_fields(component)}
-  end
-
-  defp component_fields(component) do
-    component |> Enum.map(&component_field/1) |> Enum.reject(&is_nil/1)
-  end
-
-  @simple_fields [:name, :version, :purl, :cpe, :description]
-
-  defp component_field({field, value}) when field in @simple_fields and not is_nil(value) do
-    {field, [], [[value]]}
-  end
-
-  defp component_field({:hashes, hashes}) when is_map(hashes) do
-    {:hashes, [], Enum.map(hashes, &hash/1)}
-  end
-
-  defp component_field({:licenses, [_ | _] = licenses}) do
-    {:licenses, [], Enum.map(licenses, &license/1)}
-  end
-
-  defp component_field(_other), do: nil
-
-  defp license(name) do
-    # If the name is a recognized SPDX license ID, or if we can turn it into
-    # one, we return a bom:license with a bom:id element
-    case License.spdx_id(name) do
-      nil ->
-        {:license, [],
-         [
-           {:name, [], [[name]]}
-         ]}
-
-      id ->
-        {:license, [],
-         [
-           {:id, [], [[id]]}
-         ]}
+    case options[:format] do
+      :json -> Json.bom(components, options)
+      _ -> Xml.bom(components, options)
     end
   end
 
-  defp hash({algorithm, hash}) do
-    {:hash, [alg: algorithm], [[hash]]}
-  end
-
-  defp uuid() do
+  def uuid() do
     [
       :crypto.strong_rand_bytes(4),
       :crypto.strong_rand_bytes(2),
