+++ mix.exs
@@ -28,7 +28,8 @@ defmodule SBoM.MixProject do
   # Run "mix help deps" to learn about dependencies.
   defp deps do
     [
-      {:ex_doc, "~> 0.21", only: :dev}
+      {:ex_doc, "~> 0.21", only: :dev},
+      {:jason, "~> 1.2"}
     ]
   end
